+++ CHANGELOG.md
@@ -5,6 +5,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+
+- The file `.cargo/config` will be removed before starting the build 
+
 ## [0.6.1] - 2020-05-04
 
 ### Fixed